+++ manifests/base/agent/agent-deploy.yaml
@@ -29,11 +29,6 @@ spec:
               scheme: HTTPS
             initialDelaySeconds: 10
             periodSeconds: 20
-          env:
-            - name: NAMESPACE
-              valueFrom:
-                fieldRef:
-                  fieldPath: metadata.namespace
       securityContext:
         runAsNonRoot: true
       nodeSelector:
